staticanalysis/bot: Move taskcluster mock class in separate module
